HACKATHON/requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
django-filter
drf-yasg
gunicorn
djangorestframework-simplejwt
django-cachalot
django-rest-passwordreset
coreapi
redis
celery[redis]
bs4
telebot
requests
lxml
